backend/requirements.txt
flask
pillow
flask-cors
